Revert "feat: KPI-154 edit preview table (#70)"
This reverts commit ac1cced397b396ef3c40ceb2eefbc2cc534779ea.
