+++ index.html
@@ -15,7 +15,7 @@
     </script>
   </head>
   <body>
-    <div id="root" style="width: 100%; overflow-x: auto;"></div>
+    <div id="root"></div>
     <script type="module" src="/src/main.jsx"></script>
   </body>
 </html>
