+++ package.json
@@ -6,7 +6,7 @@
     "build": "vite build",
     "build-demo": "vite build --mode demo",
     "serve": "vite preview",
-    "test": "jest --env=jsdom",
+    "test": "jest",
     "coverage": "vitest --coverage",
     "clear-jest": "jest --clearCache"
   },
@@ -27,7 +27,7 @@
     "chartjs-plugin-zoom": "1.2.0",
     "faker": "5.5.3",
     "file-saver": "^2.0.5",
-    "moment": "2.29.3",
+    "moment": "2.29.1",
     "papaparse": "^5.3.2",
     "react": "17.0.2",
     "react-chartjs-2": "4.0.0",
@@ -41,8 +41,6 @@
     "@babel/core": "^7.16.5",
     "@babel/preset-env": "^7.16.5",
     "@babel/preset-react": "^7.16.5",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
     "@vitejs/plugin-react": "^1.0.0",
     "babel-jest": "^27.4.5",
     "c8": "7.10.0",
