+++ src/service/uploadFileData.js
@@ -13,13 +13,3 @@ export const uploadFileData = async (body) => {
   const data = await response.data
   return data
 }
-
-export const validateData = async (body) => {
-  const headers = await getAuthorizationHeader()
-  const response = await axios.post(`${uploadFile}/validate_data`, body, {
-    headers: headers
-  })
-
-  const data = await response.data
-  return data
-}
