--- src/utils/constants/CompanyDescription.js
@@ -1,49 +1,0 @@
-export const VERTICALS = [
-  'Accounting & Auditing',
-  'Banking',
-  'Business Services',
-  'Consumer Services',
-  'Ecommerce',
-  'Education',
-  'Energy',
-  'Engineering',
-  'Entertainment',
-  'Environment',
-  'Fashion',
-  'Goverment',
-  'Healthcare',
-  'Heavy Industry',
-  'Hotels',
-  'Infrastructure',
-  'Insurance',
-  'Law',
-  'Life Sciences',
-  'Logistics',
-  'Manufacturing',
-  'Media',
-  'Mining',
-  'Non profits',
-  'Professional Services',
-  'Publishing',
-  'Real Estate',
-  'Transportation',
-  'Utilities',
-  'Wholesale',
-  'Horizontal'
-]
-
-export const SECTORS = [
-  'Application Software',
-  'Communication Equipment',
-  'Computer Hardware',
-  'Online Media',
-  'Semiconductors'
-]
-
-export const INVESTOR_PROFILES = [
-  'Early stage VC',
-  'Growth stage VC',
-  'Private equity',
-  'Founder backed',
-  'Family office'
-]
