+++ src/views/UploadFileView/Components/DragAndDrop.jsx
@@ -1,7 +1,13 @@
-import React from 'react'
-import { Paper, Typography, makeStyles, Box, Button } from '@material-ui/core'
-import FolderIcon from '../../../components/Icons/FolderIcon'
+import React, { useState } from 'react'
+import Papa from 'papaparse'
+import { Button, Paper, Typography, makeStyles, Snackbar, Box } from '@material-ui/core'
+import { Alert } from '@mui/material'
+import { useDropzone } from 'react-dropzone'
+import { uploadFileData } from '../../../service/uploadFileData'
+import { getUserId } from './../../../service/session'
 import RestartAltIcon from '@mui/icons-material/RestartAlt'
+import FolderIcon from '../../../components/Icons/FolderIcon'
+import PreviewTable from './PreviewTable'
 
 const useStyles = makeStyles({
   root: {
@@ -50,38 +56,93 @@ const useStyles = makeStyles({
   }
 })
 
-export function ButtonOptions ({ onCancel, setEdit, onValidateData }) {
+export default function DragAndDrop (props) {
   const classes = useStyles()
-  return (
-    <Box className={classes.buttonContainer}>
-      <Button
-        variant='outlined'
-        className={classes.resetButton}
-        onClick={onCancel}
-      >Reset
-        <RestartAltIcon color="action" fontSize="small" sx={{ marginLeft: 0.4 }}></RestartAltIcon>
-      </Button>
-      <Button
-        variant="contained"
-        className={classes.editButton}
-        onClick={() => setEdit(true)}
-      >
-        Edit
-      </Button>
-      <Button
-        variant="contained"
-        onClick={onValidateData}
-        className={classes.uploadButton}
-      >
-        Upload File
-      </Button>
-    </Box>
-  )
-}
+  const [open, setOpen] = useState(false)
+  const { onConnectETL, onDisconnectETL, onSendRegister } = props
+  const [confirmMessage, setConfirmMessage] = useState('')
+  const [headRows, setHeadRows] = useState([])
+  const [bodyRows, setBodyRows] = useState([])
+  const onDrop = (acceptedFiles, fileRejections) => {
+    if (fileRejections.length === 0) parseFile(acceptedFiles)
+  }
+  const { acceptedFiles, fileRejections, getRootProps, getInputProps } =
+    useDropzone({
+      onDrop,
+      maxFiles: 1,
+      accept: '.csv'
+    })
 
-export function DragAndDrop (props) {
-  const classes = useStyles()
-  const { getRootProps, getInputProps, acceptedFiles, acceptedFilesItems, fileRejectionItems } = props
+  const acceptedFilesItems = acceptedFiles.map((file) => (<Alert severity="success" key={file.path}>{file.path} - {file.size} bytes </Alert>))
+
+  const fileRejectionItems = fileRejections.map(({ file, errors }) => (
+    errors.map(e => (
+      <Alert severity="error" key={e.code}>{file.path}: {e.message} </Alert>
+    ))
+  ))
+
+  const parseFile = (acceptedFiles) => {
+    Papa.parse(acceptedFiles[0], {
+      complete: function (results) {
+        mapParsedData(results.data)
+      }
+    })
+  }
+
+  const mapParsedData = (parsedData) => {
+    setHeadRows(parsedData.slice(0, 3))
+    setBodyRows(parsedData.slice(3))
+  }
+
+  function getBinaryFromFile (file) {
+    return new Promise((resolve, reject) => {
+      const reader = new FileReader()
+      reader.addEventListener('load', () => resolve(reader.result))
+      reader.addEventListener('error', err => reject(err))
+      reader.readAsBinaryString(file)
+    })
+  }
+
+  const handleClose = (event, reason) => {
+    if (reason === 'clickaway') {
+      return
+    }
+
+    setOpen(false)
+  }
+
+  const callWebsocketRegister = async (user, file) => {
+    onSendRegister(user, file)
+  }
+
+  const onFileSent = async (response) => {
+    if (response.uploaded) {
+      setConfirmMessage('File was uploaded successfully!!')
+      const user = await getUserId()
+      callWebsocketRegister(user, response.filename)
+    }
+    if (response.error) {
+      onDisconnectETL()
+      setConfirmMessage(`File was not uploaded: ${response.error}`)
+    }
+    setOpen(true)
+  }
+
+  const onClick = async (_) => {
+    const fileToUpload = acceptedFiles[0]
+
+    if (fileToUpload == null) {
+      setConfirmMessage('Please, select a file to uploaded')
+      setOpen(true)
+    } else {
+      onConnectETL()
+      const response = await uploadFileData({
+        fileName: fileToUpload.name,
+        file: await (getBinaryFromFile(fileToUpload))
+      })
+      onFileSent(response)
+    }
+  }
 
   return (
     <Box>
@@ -107,6 +168,41 @@ export function DragAndDrop (props) {
           }
         </>
       </Box>
+      {acceptedFiles.length > 0 && fileRejectionItems.length === 0 && (
+        <>
+          <Box className={classes.buttonContainer}>
+            <Button
+              variant='outlined'
+              className={classes.resetButton}
+            >Reset
+              <RestartAltIcon color="action" fontSize="small" sx={{ marginLeft: 0.4 }}></RestartAltIcon>
+            </Button>
+            <Button
+              variant="contained"
+              className={classes.editButton}
+            >
+              Edit
+            </Button>
+            <Button
+              variant="contained"
+              onClick={onClick}
+              className={classes.uploadButton}
+            >
+              Upload File
+            </Button>
+          </Box>
+          <Box style={{ marginTop: '20px' }}>
+            <PreviewTable head={headRows} body={bodyRows}></PreviewTable>
+          </Box>
+        </>
+      )}
+      <Snackbar
+        anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
+        open={open}
+        autoHideDuration={2000}
+        message={confirmMessage}
+        onClose={handleClose}
+      />
     </Box>
   )
 }
