+++ src/views/UserPanelView/UserDetailView.jsx
@@ -13,8 +13,11 @@ import { PermissionsView } from './../PermissionsView/PermissionsView'
 
 const useStyles = makeStyles(theme => ({
   root: {
-    [theme.breakpoints.up('sm')]: {
-      width: '600px'
+    [theme.breakpoints.up('md')]: {
+      width: '80vh'
+    },
+    [theme.breakpoints.between('sm', 'md')]: {
+      mimWidth: '55vh'
     },
     [theme.breakpoints.down('sm')]: {
       width: '100%'
